working supplier + category add/update/delete forms

+++ src/app/categories/categories.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit, ViewChild } from '@angular/core';
 import { Category } from '../category';
 import { CategoryService } from '../category.service';
 import { MatTableDataSource, MatSort, MatPaginator } from '@angular/material';
+import { DelDialogueService } from '../del-dialogue.service';
+import { FormBuilder } from '@angular/forms';
 
 @Component({
   selector: 'app-categories',
@@ -11,13 +13,23 @@ import { MatTableDataSource, MatSort, MatPaginator } from '@angular/material';
 export class CategoriesComponent implements OnInit {
 
   categories: Category[];
+  category: Category;
+
+  updateCategoryForm = this.formBuilder.group({
+    categoryName: [''],
+  })
+
   categoryDataSource = new MatTableDataSource(this.categories);
   displayedColumns = ['categoryId', 'categoryName', 'actions'];
   
   @ViewChild(MatSort, { static: true }) sort: MatSort;
   @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator;
   
-  constructor(private categoryService: CategoryService) {
+  constructor(
+    private categoryService: CategoryService,
+    private delDialogueService: DelDialogueService,
+    private formBuilder: FormBuilder
+    ) {
     this.categoryService.findAll().subscribe(data => {
       this.categories = data;
       this.categoryDataSource.data = this.categories;
@@ -29,25 +41,41 @@ export class CategoriesComponent implements OnInit {
     this.categoryDataSource.paginator = this.paginator;
   }
 
-  public deleteCategory(categoryId: number) {
-    var category = this.categories.filter(function (category) {
-      return category.categoryId === categoryId;
+  openCatAdd() {
+    this.delDialogueService.openCatAdd().afterClosed().subscribe(add => {
+      if (add) {
+        console.log("Category added!")
+      }
     })
+  }
 
-    let categoryString = JSON.stringify(category, null, 4);
+  updateCategory(formCategory: Category) {
+    this.category = formCategory;
+    this.updateCategoryForm.patchValue({
+      categoryName: formCategory.categoryName
+    })
+  }
+
+  submitUpdate() { 
+
+    this.category.categoryName = this.updateCategoryForm.value.categoryName;
+
+    console.log(this.category);
+    this.categoryService.update(this.category, this.category.categoryId).subscribe(result => {
+      console.log("Category updated!")
+      location.reload();
+    })
+  }
+
+  deleteCategory(categoryId: number) {
     
-    this.categoryService.delete(categoryId).subscribe(result => {
-      this.categoryService.findAll().subscribe(data => {
-        this.categories = data;
-        this.categoryDataSource.data = this.categories;
-      })
+    this.delDialogueService.openConfirm('WARNING: Removing this category will remove all associated products.').afterClosed().subscribe(remove => {
+      if (remove) {
+        this.categoryService.delete(categoryId).subscribe(result => {
+          location.reload();
+        })
+      }
     })
-
-    console.log("Category and associated products deleted.");
-
-    error => {
-      console.error("Category delete function error\n" + error);
-    }
   }
 
 headerTitle = "Product Categories List"
