+++ src/app/suppliers/suppliers.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit, ViewChild } from '@angular/core';
 import { SupplierService } from '../supplier.service';
 import { Supplier } from '../supplier';
 import { MatTableDataSource, MatSort, MatPaginator } from '@angular/material';
+import { DelDialogueService } from '../del-dialogue.service';
+import { FormBuilder } from '@angular/forms';
 
 @Component({
   selector: 'app-suppliers',
@@ -11,13 +13,23 @@ import { MatTableDataSource, MatSort, MatPaginator } from '@angular/material';
 export class SuppliersComponent implements OnInit {
 
   suppliers: Supplier[];
+  supplier: Supplier;
+
+  updateSupplierForm = this.formBuilder.group({
+    supplierName: [''],
+  })
+
   supplierDataSource = new MatTableDataSource(this.suppliers);
   displayedColumns = ['supplierId', 'supplierName', 'actions'];
 
   @ViewChild(MatSort, { static: true }) sort: MatSort;
   @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator;
 
-  constructor(private supplierService: SupplierService) {
+  constructor(
+    private supplierService: SupplierService,
+    private delDialogueService: DelDialogueService,
+    private formBuilder: FormBuilder
+  ) {
     this.supplierService.findAll().subscribe(data => {
       this.suppliers = data;
       this.supplierDataSource.data = this.suppliers;
@@ -29,27 +41,42 @@ export class SuppliersComponent implements OnInit {
     this.supplierDataSource.paginator = this.paginator;
   }
 
-  public deleteSupplier(supplierId: number) {
-    var supplier = this.suppliers.filter(function (supplier) {
-      return supplier.supplierId === supplierId;
+  openSupAdd() {
+    this.delDialogueService.openSupAdd().afterClosed().subscribe(add => {
+      if (add) {
+        console.log("Supplier added!")
+      }
     })
+  }
 
-    let supplierString = JSON.stringify(supplier, null, 4);
-
-    confirm("WARNING: Deleting this supplier will remove all of its products!\n" + supplierString);
-
-    this.supplierService.delete(supplierId).subscribe(result => {
-      this.supplierService.findAll().subscribe(data => {
-        this.suppliers = data;
-        this.supplierDataSource.data = this.suppliers;
-      })
+  updateSupplier(formSupplier: Supplier) {
+    this.supplier = formSupplier;
+    this.updateSupplierForm.patchValue({
+      supplierName: formSupplier.supplierName
     })
+  }
 
-    console.log("Supplier and associated products deleted.")
+  submitUpdate() {
 
-    error => {
-      console.error("Supplier delete function error\n" + error);
-    }
+    this.supplier.supplierName = this.updateSupplierForm.value.supplierName;
+
+    console.log(this.supplier);
+
+    this.supplierService.update(this.supplier, this.supplier.supplierId).subscribe(result => {
+      console.log("Supplier updated!")
+      location.reload();
+    })
+  }
+
+  deleteSupplier(supplierId: number) {
+
+    this.delDialogueService.openConfirm('WARNING: Removing this supplier will remove all associated products.').afterClosed().subscribe(remove => {
+      if (remove) {
+        this.supplierService.delete(supplierId).subscribe(result => {
+          location.reload();
+        })
+      }
+    })
   }
 
   headerTitle = "Product Suppliers List"
